requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.5
et-xmlfile==1.1.0
Flask==3.0.3
flask-restx==1.3.0
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==2.0.1
openpyxl==3.1.5
panda==0.3.1
pandas==2.2.2
prompt_toolkit==3.0.47
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywinauto==0.6.8
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
six==1.16.0
SQLAlchemy==2.0.31
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
